API

LAMMPS.LMP — Type

LMP(f::Function, args=String[], comm=nothing)

Create a new LAMMPS instance and call f on that instance while returning the result from f.

LAMMPS.LMP — Type

LMP(args::Vector{String}=String[], comm::Union{Nothing, MPI.Comm}=nothing)

Create a new LAMMPS instance while passing in a list of strings as if they were command-line arguments for the LAMMPS executable.

For a full ist of Command-line options see: https://docs.lammps.org/Run_options.html

LAMMPS.close! — Method

close!(lmp::LMP)

Shutdown a LAMMPS instance.

LAMMPS.command — Method

command(lmp::LMP, cmd::Union{String, Array{String}})

Process LAMMPS input commands from a String or from an Array of Strings.

For a full list of commands see: https://docs.lammps.org/commands_list.html

This function processes a multi-line string similar to a block of commands from a file. The string may have multiple lines (separated by newline characters) and also single commands may be distributed over multiple lines with continuation characters (’&’). Those lines are combined by removing the ‘&’ and the following newline character. After this processing the string is handed to LAMMPS for parsing and executing.

Arrays of Strings get concatenated into a single String inserting newline characters as needed.

LAMMPS.jl 0.4.1

Multiline string support """ and support for array of strings was added. Prior versions of LAMMPS.jl ignore newline characters.

Examples

LMP(["-screen", "none"]) do lmp
    command(lmp, """
        atom_modify map yes
        region cell block 0 2 0 2 0 2
        create_box 1 cell
        lattice sc 1
        create_atoms 1 region cell
        mass 1 1

        group a id 1 2 3 5 8
        group even id 2 4 6 8
        group odd id 1 3 5 7
    """)
end

LAMMPS.extract_atom — Function

extract_atom(lmp, name, dtype=nothing, axes1, axes2)

LAMMPS.extract_compute — Method

extract_compute(lmp, name, style, type)

LAMMPS.extract_global — Function

extract_global(lmp, name, dtype=nothing)

LAMMPS.extract_variable — Function

extract_variable(lmp::LMP, name, group)

Extracts the data from a LAMMPS variable. When the variable is either an equal-style compatible variable, a vector-style variable, or an atom-style variable, the variable is evaluated and the corresponding value(s) returned. Variables of style internal are compatible with equal-style variables, if they return a numeric value. For other variable styles, their string value is returned.

LAMMPS.gather — Function

gather(lmp::LMP, name::String, T::Union{Type{Int32}, Type{Float64}}, ids::Union{Nothing, Array{Int32}}=nothing)

Gather the named per-atom, per-atom fix, per-atom compute, or fix property/atom-based entities from all processes. By default (when ids=nothing), this method collects data from all atoms in consecutive order according to their IDs. The optional parameter ids determines for which subset of atoms the requested data will be gathered. The returned data will then be ordered according to ids

Compute entities have the prefix c_, fix entities use the prefix f_, and per-atom entites have no prefix.

The returned Array is decoupled from the internal state of the LAMMPS instance.

Type Verification

Due to how the underlying C-API works, it's not possible to verify the element data-type of fix or compute style data. Supplying the wrong data-type will not throw an error but will result in nonsensical output

ids

The optional parameter ids only works, if there is a map defined. For example by doing: command(lmp, "atom_modify map yes") However, LAMMPS only issues a warning if that's the case, which unfortuately cannot be detected through the underlying API. Starting form LAMMPS version 17 Apr 2024 this should no longer be an issue, as LAMMPS then throws an error instead of a warning.

LAMMPS.get_category_ids — Function

get_category_ids(lmp::LMP, category::String, buffer_size::Integer=50)

Look up the names of entities within a certain category.

Valid categories are: compute, dump, fix, group, molecule, region, and variable. names longer than buffer_size will be truncated to fit inside the buffer.

LAMMPS.get_natoms — Method

get_natoms(lmp::LMP)::Int64

Get the total number of atoms in the LAMMPS instance.

Will be precise up to 53-bit signed integer due to the underlying lammps_get_natoms returning a Float64.

LAMMPS.group_to_atom_ids — Method

group_to_atom_ids(lmp::LMP, group::String)

Find the IDs of the Atoms in the group.

LAMMPS.locate — Method

locate()

Locate the LAMMPS library currently being used, by LAMMPS.jl

LAMMPS.scatter! — Method

scatter!(lmp::LMP, name::String, data::VecOrMat{T}, ids::Union{Nothing, Array{Int32}}=nothing) where T<:Union{Int32, Float64}

Scatter the named per-atom, per-atom fix, per-atom compute, or fix property/atom-based entity in data to all processes. By default (when ids=nothing), this method scatters data to all atoms in consecutive order according to their IDs. The optional parameter ids determines to which subset of atoms the data will be scattered.

Compute entities have the prefix c_, fix entities use the prefix f_, and per-atom entites have no prefix.

Type Verification

Due to how the underlying C-API works, it's not possible to verify the element data-type of fix or compute style data. Supplying the wrong data-type will not throw an error but will result in nonsensical date being supplied to the LAMMPS instance.

ids

The optional parameter ids only works, if there is a map defined. For example by doing: command(lmp, "atom_modify map yes") However, LAMMPS only issues a warning if that's the case, which unfortuately cannot be detected through the underlying API. Starting form LAMMPS version 17 Apr 2024 this should no longer be an issue, as LAMMPS then throws an error instead of a warning.

LAMMPS.set_library! — Method

set_library!(path)

Change the library path used by LAMMPS.jl for liblammps.so to path.

Note

You will need to restart Julia to use the new library.

Warning

Due to a bug in Julia (until 1.6.5 and 1.7.1), setting preferences in transitive dependencies is broken (https://github.com/JuliaPackaging/Preferences.jl/issues/24). To fix this either update your version of Julia, or add LAMMPS_jll as a direct dependency to your project.