API

LAMMPS.close! — Method

close!(lmp::LMP)

Shutdown an LMP instance.

LAMMPS.command — Method

command(lmp::lmp, cmd)

LAMMPS.extract_atom — Function

extract_atom(lmp, name, dtype=nothing, axes1, axes2)

LAMMPS.extract_compute — Method

extract_compute(lmp, name, style, type)

LAMMPS.extract_global — Function

extract_global(lmp, name, dtype=nothing)

LAMMPS.extract_variable — Function

extract_variable(lmp::LMP, name, group)

Extracts the data from a LAMMPS variable. When the variable is either an equal-style compatible variable, a vector-style variable, or an atom-style variable, the variable is evaluated and the corresponding value(s) returned. Variables of style internal are compatible with equal-style variables, if they return a numeric value. For other variable styles, their string value is returned.

LAMMPS.gather — Function

gather(lmp::LMP, name::String, T::Union{Type{Int32}, Type{Float64}}, ids::Union{Nothing, Array{Int32}}=nothing)

Gather the named per-atom, per-atom fix, per-atom compute, or fix property/atom-based entities from all processes. By default (when ids=nothing), this method collects data from all atoms in consecutive order according to their IDs. The optional parameter ids determines for which subset of atoms the requested data will be gathered. The returned data will then be ordered according to ids

Compute entities have the prefix c_, fix entities use the prefix f_, and per-atom entites have no prefix.

The returned Array is decoupled from the internal state of the LAMMPS instance.

Type Verification

Due to how the underlying C-API works, it's not possible to verify the element data-type of fix or compute style data. Supplying the wrong data-type will not throw an error but will result in nonsensical output

ids

The optional parameter ids only works, if there is a map defined. For example by doing: command(lmp, "atom_modify map yes") However, LAMMPS only issues a warning if that's the case, which unfortuately cannot be detected through the underlying API. Starting form LAMMPS version 17 Apr 2024 this should no longer be an issue, as LAMMPS then throws an error instead of a warning.

LAMMPS.get_natoms — Method

get_natoms(lmp::LMP)::Int64

Get the total number of atoms in the LAMMPS instance.

Will be precise up to 53-bit signed integer due to the underlying lammps_get_natoms returning a Float64.

LAMMPS.locate — Method

locate()

Locate the LAMMPS library currently being used, by LAMMPS.jl

LAMMPS.scatter! — Method

scatter!(lmp::LMP, name::String, data::VecOrMat{T}, ids::Union{Nothing, Array{Int32}}=nothing) where T<:Union{Int32, Float64}

Scatter the named per-atom, per-atom fix, per-atom compute, or fix property/atom-based entity in data to all processes. By default (when ids=nothing), this method scatters data to all atoms in consecutive order according to their IDs. The optional parameter ids determines to which subset of atoms the data will be scattered.

Compute entities have the prefix c_, fix entities use the prefix f_, and per-atom entites have no prefix.

Type Verification

Due to how the underlying C-API works, it's not possible to verify the element data-type of fix or compute style data. Supplying the wrong data-type will not throw an error but will result in nonsensical date being supplied to the LAMMPS instance.

ids

The optional parameter ids only works, if there is a map defined. For example by doing: command(lmp, "atom_modify map yes") However, LAMMPS only issues a warning if that's the case, which unfortuately cannot be detected through the underlying API. Starting form LAMMPS version 17 Apr 2024 this should no longer be an issue, as LAMMPS then throws an error instead of a warning.

LAMMPS.set_library! — Method

set_library!(path)

Change the library path used by LAMMPS.jl for liblammps.so to path.

Note

You will need to restart Julia to use the new library.

Warning

Due to a bug in Julia (until 1.6.5 and 1.7.1), setting preferences in transitive dependencies is broken (https://github.com/JuliaPackaging/Preferences.jl/issues/24). To fix this either update your version of Julia, or add LAMMPS_jll as a direct dependency to your project.