API

LAMMPS.extract_atom — Function

extract_atom(lmp, name, dtype=nothing, )

LAMMPS.get_natoms — Method

get_natoms(lmp::LMP)::Int64

Get the total number of atoms in the LAMMPS instance.

Will be precise up to 53-bit signed integer due to the underlying lammps_get_natoms returning a Float64.

LAMMPS.locate — Method

locate()

Locate the LAMMPS library currently being used, by LAMMPS.jl

LAMMPS.neighbors — Method

neighbors(lmb::LMP, idx, element)

Given a neighbor list idx and the element therein, return the atom index, and it's neigbors.

LAMMPS.set_library! — Method

set_library!(path)

Change the library path used by LAMMPS.jl for liblammps.so to path.

Note

You will need to restart Julia to use the new library.

Warning

Due to a bug in Julia (until 1.6.5 and 1.7.1), setting preferences in transitive dependencies is broken (https://github.com/JuliaPackaging/Preferences.jl/issues/24). To fix this either update your version of Julia, or add LAMMPS_jll as a direct dependency to your project.